LayoutTests/fast/element-targeting/target-rendered-text-excludes-live-text-expected.txt


This test requires WebKitTestRunner

PASS text is "{320,240}"
PASS successfullyParsed is true

TEST COMPLETE
